[MIG][11.0] Migration of mozaik_email into 11.0 + Flake/pylint green + Edit 'abstract' unit test in mozaik_coordinate
