+++ odoo/addons/inherit_abstract_view/__init__.py
@@ -1,0 +1,1 @@
+from . import models
