+++ odoo/addons/inherit_abstract_view/__manifest__.py
@@ -1,20 +1,21 @@
+# -*- coding: utf-8 -*-
 # Copyright 2018 ACSONE SA/NV
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 {
-    'name': 'Mozaik: Base Abstract Model',
+    'name': 'Inherit Abstract View',
     'description': """
-        Abstract model for mozaik models""",
+        Allow a view to inherit a abstract view
+        (defined for a abstract model)""",
     'version': '11.0.1.0.0',
     'license': 'AGPL-3',
     'author': 'ACSONE SA/NV',
     'website': 'https://acsone.eu/',
-    'category': 'Political Association',
     'depends': [
         'base',
-        'mail',
-        'inherit_abstract_view',
     ],
     'data': [
-        'security/res_groups.xml',
+    ],
+    'demo': [
     ],
 }
