+++ odoo/addons/inherit_abstract_view/models/__init__.py
@@ -1,0 +1,1 @@
+from . import ir_ui_view
