+++ odoo/addons/mozaik_coordinate/tests/__init__.py
@@ -1,5 +1,9 @@
 # Copyright 2018 ACSONE SA/NV (<http://acsone.eu>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from . import common
+from . import common_abstract_coordinate
+from . import common_coordinate_wizard
+from . import common_failure_editor
 from . import test_abstract_coordinate
 from . import test_failure_editor
 from . import test_coordinate_wizard
