--- odoo/addons/mozaik_email/data/email_data.xml
@@ -1,15 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <!-- Bounce Notification -->
-
-        <record id="email_failure_notification" model="mail.message.subtype">
-            <field name="name">Email Failure Notification</field>
-            <field name="res_model">email.coordinate</field>
-            <field name="default" eval="False"/>
-            <field name="description">This Email Coordinate seems to be down</field>
-        </record>
-
-    </data>
-</openerp>
