+++ odoo/addons/mozaik_email/demo/email_coordinate.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+    <record model="email.coordinate" id="email_coordinate_admin">
+        <field name="email">admin@example.test</field>
+        <field name="partner_id" ref="base.partner_root"/>
+        <field name="is_main" eval="True"/>
+    </record>
+
+    <record model="email.coordinate" id="email_coordinate_thierry">
+        <field name="email">thierry@example.test</field>
+        <field name="partner_id" ref="mozaik_coordinate.res_partner_thierry"/>
+        <field name="is_main" eval="True"/>
+    </record>
+
+    <record model="email.coordinate" id="email_coordinate_pauline">
+        <field name="email">pauline@example.test</field>
+        <field name="partner_id" ref="mozaik_coordinate.res_partner_pauline"/>
+        <field name="is_main" eval="True"/>
+    </record>
+
+    <record model="email.coordinate" id="email_coordinate_nicolas">
+        <field name="email">nicolas@example.test</field>
+        <field name="partner_id" ref="mozaik_coordinate.res_partner_nicolas"/>
+        <field name="is_main" eval="True"/>
+    </record>
+
+    <record model="email.coordinate" id="email_coordinate_sandra">
+        <field name="email">sandra@example.test</field>
+        <field name="partner_id" ref="mozaik_coordinate.res_partner_sandra"/>
+        <field name="is_main" eval="True"/>
+    </record>
+</odoo>
