+++ odoo/addons/mozaik_email/models/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2018 ACSONE SA/NV (<http://acsone.eu>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import models
-from . import wizards
+from . import email_coordinate
+from . import res_partner
