--- odoo/addons/mozaik_email/res_partner_view.xml
@@ -1,73 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <record id="view_partner_tree" model="ir.ui.view">
-            <field name="name">res.partner.tree (mozaik_email)</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_tree" />
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='email']" position="attributes">
-                    <attribute name="invisible">1</attribute>
-                </xpath>
-            </field>
-        </record>
-
-        <record id="view_partner_form" model="ir.ui.view">
-            <field name="name">res.partner.form (mozaik_email)</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form" />
-            <field name="arch" type="xml">
-
-                <xpath expr="//field[@name='email']" position="replace">
-                    <field name="email_coordinate_id" />
-                </xpath>
-
-            </field>
-        </record>
-
-        <record id="view_partner_2_form" model="ir.ui.view">
-            <field name="name">res.partner.form (mozaik_email_2)</field>
-            <field name="model">res.partner</field>
-            <field name="inherit_id" ref="mozaik_coordinate.view_partner_form" />
-            <field name="arch" type="xml">
-
-                <xpath expr="//separator[@name='email']" position="replace">
-                    <span name="email" class="oe_horizontal_separator oe_clear oe_inline">
-                        Email Coordinates
-                    </span>
-                    <button class="oe_right" string="All Details" name="%(email_coordinate_action)d"
-                            type="action" attrs="{'invisible': [('id','=',False)]}"
-                            context="{'search_default_partner_id': active_id,
-                                      'default_partner_id': active_id,
-                                      'search_default_all': True}" />
-                    <field name="email_coordinate_ids" context="{'default_partner_id': active_id}"
-                           attrs="{'invisible': [('active','=',False)], 'readonly': [('id','=',False)]}">
-                        <tree colors="darkorange:is_duplicate_detected==True;red:unauthorized==True">
-                            <field name="email"/>
-                            <field name="coordinate_category_id" />
-                            <field name="is_main" />
-                            <field name="unauthorized" />
-                            <field name="vip" groups="mozaik_base.mozaik_res_groups_vip_reader" />
-                            <field name="bounce_counter" attrs="{'invisible': [('bounce_counter','=',0)]}" />
-                            <field name="is_duplicate_detected" invisible="1"/>
-                        </tree>
-                    </field>
-                    <field name="email_coordinate_inactive_ids"
-                           readonly="1" attrs="{'invisible': [('active','=',True)]}">
-                        <tree colors="red:unauthorized==True">
-                            <field name="email"/>
-                            <field name="coordinate_category_id" />
-                            <field name="is_main" />
-                            <field name="unauthorized" />
-                            <field name="vip" groups="mozaik_base.mozaik_res_groups_vip_reader" />
-                            <field name="bounce_counter" attrs="{'invisible': [('bounce_counter','=',0)]}" />
-                        </tree>
-                    </field>
-                </xpath>
-
-            </field>
-        </record>
-
-    </data>
-</openerp>
