+++ odoo/addons/mozaik_email/security/abstract_coordinate.xml
@@ -1,0 +1,93 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record model="ir.model.access" id="email_coordinate_access_user_r">
+        <field name="name">email.coordinate all R</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="group_id" ref="mozaik_coordinate.res_groups_coordinate_user_reader"/>
+        <field name="perm_read" eval="1"/>
+        <field name="perm_create" eval="0"/>
+        <field name="perm_write" eval="0"/>
+        <field name="perm_unlink" eval="0"/>
+    </record>
+
+    <record model="ir.model.access" id="email_coordinate_access_user_cud">
+        <field name="name">email.coordinate user CUD</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="group_id" ref="mozaik_coordinate.res_groups_coordinate_user"/>
+        <field name="perm_read" eval="0"/>
+        <field name="perm_create" eval="1"/>
+        <field name="perm_write" eval="1"/>
+        <field name="perm_unlink" eval="1"/>
+    </record>
+
+    <record model="ir.model.access" id="email_coordinate_access_vip_r">
+        <field name="name">email.coordinate vip R</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="group_id" ref="mozaik_coordinate.res_groups_coordinate_vip_reader"/>
+        <field name="perm_read" eval="1"/>
+        <field name="perm_create" eval="0"/>
+        <field name="perm_write" eval="0"/>
+        <field name="perm_unlink" eval="0"/>
+    </record>
+
+    <record model="ir.model.access" id="email_coordinate_access_vip_cud">
+        <field name="name">email.coordinate vip CUD</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="group_id" ref="mozaik_coordinate.res_groups_coordinate_vip_user"/>
+        <field name="perm_read" eval="0"/>
+        <field name="perm_create" eval="1"/>
+        <field name="perm_write" eval="1"/>
+        <field name="perm_unlink" eval="1"/>
+    </record>
+
+    <record id="email_coordinate_reader_rule" model="ir.rule">
+        <field name="name">Email Coordinate for Mozaik Reader/User</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="domain_force">[('vip','=',False)]</field>
+        <field name="groups" eval="[(6, 0, [
+            ref('mozaik_coordinate.res_groups_coordinate_user_reader'),
+            ref('mozaik_coordinate.res_groups_coordinate_user'),
+        ])]"/>
+        <field name="perm_read" eval="1"/>
+        <field name="perm_write" eval="0"/>
+        <field name="perm_create" eval="0"/>
+        <field name="perm_unlink" eval="0"/>
+    </record>
+
+    <record id="email_coordinate_vip_reader_rule" model="ir.rule">
+        <field name="name">Email Coordinate for Mozaik VIP Reader</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="domain_force">[(1,'=',1)]</field>
+        <field name="groups" eval="[(6, 0, [ref('mozaik_coordinate.res_groups_coordinate_vip_reader')])]"/>
+        <field name="perm_read" eval="1"/>
+        <field name="perm_write" eval="0"/>
+        <field name="perm_create" eval="0"/>
+        <field name="perm_unlink" eval="0"/>
+    </record>
+
+    <record id="email_coordinate_user_rule" model="ir.rule">
+        <field name="name">Email Coordinate for Mozaik User</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="domain_force">[('vip','=',False)]</field>
+        <field name="groups" eval="[(6, 0, [ref('mozaik_coordinate.res_groups_coordinate_user')])]"/>
+        <field name="perm_read" eval="0"/>
+        <field name="perm_write" eval="1"/>
+        <field name="perm_create" eval="1"/>
+        <field name="perm_unlink" eval="1"/>
+    </record>
+
+    <record id="email_coordinate_manager_rule" model="ir.rule">
+        <field name="name">Email Coordinate for Mozaik Manager</field>
+        <field name="model_id" ref="model_email_coordinate"/>
+        <field name="domain_force">[(1,'=',1)]</field>
+        <field name="groups" eval="[(6, 0, [ref('mozaik_coordinate.res_groups_coordinate_manager')])]"/>
+        <field name="perm_read" eval="0"/>
+        <field name="perm_write" eval="1"/>
+        <field name="perm_create" eval="1"/>
+        <field name="perm_unlink" eval="1"/>
+    </record>
+
+</odoo>
