--- odoo/addons/mozaik_email/security/email_security.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="0">
-
-        <record id="email_coordinate_reader_rule" model="ir.rule">
-            <field name="name">Email Coordinate for Mozaik Reader/User</field>
-            <field name="model_id" ref="model_email_coordinate"/>
-            <field name="domain_force">[('vip','=',False)]</field>
-            <field name="groups" eval="[(6, 0, [
-                ref('mozaik_base.mozaik_res_groups_reader'),
-                ref('mozaik_base.mozaik_res_groups_user'),
-            ])]"/>
-            <field name="perm_read" eval="1"/><field name="perm_write" eval="0"/>
-            <field name="perm_create" eval="0"/><field name="perm_unlink" eval="0"/>
-        </record>
-
-        <record id="email_coordinate_vip_reader_rule" model="ir.rule">
-            <field name="name">Email Coordinate for Mozaik VIP Reader</field>
-            <field name="model_id" ref="model_email_coordinate"/>
-            <field name="domain_force">[(1,'=',1)]</field>
-            <field name="groups" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_vip_reader')])]"/>
-            <field name="perm_read" eval="1"/><field name="perm_write" eval="0"/>
-            <field name="perm_create" eval="0"/><field name="perm_unlink" eval="0"/>
-        </record>
-
-        <record id="email_coordinate_user_rule" model="ir.rule">
-            <field name="name">Email Coordinate for Mozaik User</field>
-            <field name="model_id" ref="model_email_coordinate"/>
-            <field name="domain_force">[('vip','=',False)]</field>
-            <field name="groups" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_user')])]"/>
-            <field name="perm_read" eval="0"/><field name="perm_write" eval="1"/>
-            <field name="perm_create" eval="1"/><field name="perm_unlink" eval="1"/>
-        </record>
-
-        <record id="email_coordinate_manager_rule" model="ir.rule">
-            <field name="name">Email Coordinate for Mozaik Manager</field>
-            <field name="model_id" ref="model_email_coordinate"/>
-            <field name="domain_force">[(1,'=',1)]</field>
-            <field name="groups" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_manager')])]"/>
-            <field name="perm_read" eval="0"/><field name="perm_write" eval="1"/>
-            <field name="perm_create" eval="1"/><field name="perm_unlink" eval="1"/>
-        </record>
-
-    </data>
-</openerp>
