--- odoo/addons/mozaik_email/security/ir.model.access.csv
@@ -1,3 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_email_coordinate_res_group_base,mozaik.email.coordinate.res.group.base,model_email_coordinate,mozaik_base.res_groups_base,1,0,0,0
-access_email_coordinate_res_group_user,mozaik.email.coordinate.res.group.user,model_email_coordinate,mozaik_base.mozaik_res_groups_user,1,1,1,1
