--- odoo/addons/mozaik_email/tests/data/email_data.xml
@@ -1,27 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <!-- Create Email Coordinates -->
-
-        <!-- Partner Pauline -->
-
-        <record id="email_coordinate_pauline" model="email.coordinate">
-            <field name="email">pauline@gmail.com</field>
-            <field name="partner_id" ref="res_partner_pauline" />
-        </record>
-
-        <!-- Partner Thierry -->
-
-        <record id="email_coordinate_thierry_one" model="email.coordinate">
-            <field name="email">thierry@gmail.com</field>
-            <field name="partner_id" ref="res_partner_thierry" />
-        </record>
-
-        <record id="email_coordinate_thierry_two" model="email.coordinate">
-            <field name="email">titi@hotmail.com</field>
-            <field name="partner_id" ref="res_partner_thierry" />
-        </record>
-
-    </data>
-</openerp>
