+++ odoo/addons/mozaik_email/views/res_partner.xml
@@ -1,0 +1,76 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record id="view_partner_tree" model="ir.ui.view">
+        <field name="name">res.partner.tree (in mozaik_email)</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_tree" />
+        <field name="priority" eval="150"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='email']" position="attributes">
+                <attribute name="invisible">1</attribute>
+            </xpath>
+        </field>
+    </record>
+
+    <record id="view_partner_form" model="ir.ui.view">
+        <field name="name">res.partner.form (in mozaik_email)</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_form" />
+        <field name="priority" eval="150"/>
+        <field name="arch" type="xml">
+
+            <xpath expr="//field[@name='email']" position="replace">
+                <field name="email_coordinate_id" />
+            </xpath>
+
+        </field>
+    </record>
+
+    <record id="view_partner_2_form" model="ir.ui.view">
+        <field name="name">res.partner.form (in mozaik_email)</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="mozaik_coordinate.res_partner_form_view" />
+        <field name="priority" eval="150"/>
+        <field name="arch" type="xml">
+
+            <xpath expr="//page[@name='coordinates']" position="inside">
+                <span name="email" class="oe_horizontal_separator oe_clear oe_inline">
+                    Email Coordinates
+                </span>
+                <button class="oe_right" string="All Details" name="%(email_coordinate_action)d"
+                        type="action" attrs="{'invisible': [('id','=',False)]}"
+                        context="{'search_default_partner_id': active_id,
+                                      'default_partner_id': active_id,
+                                      'search_default_all': True}" />
+                <field name="email_coordinate_ids" context="{'default_partner_id': active_id}"
+                       attrs="{'invisible': [('active','=',False)], 'readonly': [('id','=',False)]}">
+                    <tree decoration-warning="unauthorized==True" decoration-info="is_duplicate_detected==True">
+                        <field name="email"/>
+                        <field name="coordinate_category_id" />
+                        <field name="is_main" />
+                        <field name="unauthorized" />
+                        <field name="vip" groups="mozaik_coordinate.res_groups_coordinate_vip_reader"/>
+                        <field name="failure_counter" attrs="{'invisible': [('failure_counter','=',0)]}" />
+                        <field name="is_duplicate_detected" invisible="1"/>
+                    </tree>
+                </field>
+                <field name="email_coordinate_inactive_ids"
+                       readonly="1" attrs="{'invisible': [('active','=',True)]}">
+                    <tree decoration-warning="unauthorized==True">
+                        <field name="email"/>
+                        <field name="coordinate_category_id" />
+                        <field name="is_main" />
+                        <field name="unauthorized" />
+                        <field name="vip" groups="mozaik_coordinate.res_groups_coordinate_vip_reader"/>
+                        <field name="failure_counter" attrs="{'invisible': [('failure_counter','=',0)]}" />
+                    </tree>
+                </field>
+            </xpath>
+
+        </field>
+    </record>
+
+</odoo>
