--- odoo/addons/mozaik_email/wizard/allow_duplicate_view.xml
@@ -1,26 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <record id="allow_duplicate_action" model="ir.actions.act_window">
-            <field name="name">Allow Duplicates</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="src_model">email.coordinate</field>
-            <field name="res_model">allow.duplicate.wizard</field>
-            <field name="multi" eval="1"/>
-            <field name="target">new</field>
-            <field name="view_id" ref="mozaik_duplicate.allow_duplicate_wizard" />
-            <field name="groups_id" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_user')])]" />
-        </record>
-
-        <record id="allow_duplicate_value" model="ir.values">
-            <field name="name">Allow Duplicates</field>
-            <field name="model">email.coordinate</field>
-            <field name="key">action</field>
-            <field name="key2">client_action_multi</field>
-            <field name="value" eval="'ir.actions.act_window,%s' % ref('allow_duplicate_action')" />
-        </record>
-
-    </data>
-</openerp>
