--- odoo/addons/mozaik_email/wizard/change_main_email.xml
@@ -1,65 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <!-- Change Main Email Wizard on res.partner -->
-
-        <record id="change_main_email_wizard" model="ir.ui.view">
-            <field name="name">change.main.email.wizard</field>
-            <field name="model">change.main.email</field>
-            <field name="inherit_id" ref="mozaik_coordinate.change_main_coordinate_wizard" />
-            <field name="arch" type="xml">
-
-                <xpath expr="//field[@name='invalidate_previous_coordinate']" position="before">
-                    <field name="old_email" options='{"no_open":True}' readonly='1' attrs="{'invisible': [('old_email','=',False)]}"/>
-                    <field name="email" readonly="context.get('mode', '') == 'switch'"/>
-                </xpath>
-
-                <xpath expr="//field[@name='invalidate_previous_coordinate']" position="attributes">
-                    <attribute name="attrs" translation="off">{'invisible': [('old_email','=',False)]}</attribute>
-                </xpath>
-
-            </field>
-        </record>
-
-        <record id="change_main_email_action" model="ir.actions.act_window">
-            <field name="name">Change Main Email</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="src_model">res.partner</field>
-            <field name="res_model">change.main.email</field>
-            <field name="target">new</field>
-            <field name="view_id" ref="change_main_email_wizard"/>
-            <field name="groups_id" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_user')])]" />
-            <field name="context">{'target_model': 'email.coordinate'}</field>
-        </record>
-
-        <record id="change_main_email_value" model="ir.values">
-            <field name="name">Change Main Email</field>
-            <field name="model">res.partner</field>
-            <field name="key">action</field>
-            <field name="key2">client_action_multi</field>
-            <field name="value" eval="'ir.actions.act_window,%s' % ref('change_main_email_action')"/>
-        </record>
-        
-        <record id="set_as_main_action" model="ir.actions.act_window">
-            <field name="name">Set as main</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="src_model">email.coordinate</field>
-            <field name="res_model">change.main.email</field>
-            <field name="target">new</field>
-            <field name="view_id" ref="change_main_email_wizard" />
-            <field name="groups_id" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_user')])]" />
-            <field name="context">{'target_model': 'email.coordinate', 'mode': 'switch'}</field>
-        </record>
-
-        <record id="set_as_main_value" model="ir.values">
-            <field name="name">Set as main</field>
-            <field name="model">email.coordinate</field>
-            <field name="key">action</field>
-            <field name="key2">client_action_multi</field>
-            <field name="value" eval="'ir.actions.act_window,%s' % ref('set_as_main_action')" />
-        </record>
-    </data>
-</openerp>
