--- odoo/addons/mozaik_email/wizard/export_vcard_view.xml
@@ -1,51 +1,0 @@
-<?xml version='1.0' encoding='UTF-8' ?>
-<openerp>
-    <data noupdate="0">
-
-        <!-- Export active email coordinates to VCF -->
-
-        <record id="export_vcard_wizard" model="ir.ui.view">
-            <field name="name">export.vcard.wizard</field>
-            <field name="model">export.vcard</field>
-            <field name="arch" type="xml">
-                <form string="Export vCard" version="7.0">
-                    <div>
-                        <label string="This will export the selected coordinates to a vCard file"/>
-                    </div>
-                    <group>
-                        <field name="export_file"
-                           nolabel="1"
-                           readonly="1"
-                           filename="export_filename" />
-                    </group>
-                    <footer>
-                        <span name="go-wizard" attrs="{'invisible': [('export_file', '!=', False)]}">
-                            <button name="export" string="Export" type="object" class="oe_highlight"/>
-                            or
-                        </span>
-                        <button string="Cancel" class="oe_link" special="cancel"/>
-                    </footer>
-                </form>
-            </field>
-        </record>
-
-        <record id="export_vcard_action" model="ir.actions.act_window">
-            <field name="name">Export vCard</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="src_model">email.coordinate</field>
-            <field name="res_model">export.vcard</field>
-            <field name="target">new</field>
-            <field name="view_id" ref="export_vcard_wizard"/>
-            <field name="groups_id" eval="[(6, 0, [ref('mozaik_base.mozaik_res_groups_user')])]"/>
-        </record>
-
-        <record id="export_vcard_values" model="ir.values">
-            <field name="name">Export vCard</field>
-            <field name="model">email.coordinate</field>
-            <field name="key">action</field>
-            <field name="key2">client_action_multi</field>
-            <field name="value" eval="'ir.actions.act_window,%s' % ref('export_vcard_action')"/>
-        </record>
-    </data>
-</openerp>
