+++ odoo/addons/mozaik_email/wizards/allow_duplicate_wizard.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 ACSONE SA/NV
+     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record id="allow_duplicate_action" model="ir.actions.act_window">
+        <field name="name">Allow Duplicates</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form</field>
+        <field name="src_model">email.coordinate</field>
+        <field name="res_model">allow.duplicate.wizard</field>
+        <field name="multi" eval="1"/>
+        <field name="target">new</field>
+        <field name="view_id" ref="mozaik_duplicate.allow_duplicate_form_view" />
+        <field name="binding_model_id" ref="mozaik_email.model_email_coordinate" />
+    </record>
+
+</odoo>
